Phase 2 of the Cloudflare migration: D1 schema, FTS5 search, export from Supabase (ADR-45)
The schema is drizzle-orm/sqlite-core (enums as text with CHECK constraints keeping enumValues, JSON as text, timestamps as epoch milliseconds, booleans as 0/1; docs/schema-notes.md has every rule). Full text is an FTS5 external-content table with column weights that mirror the old vector, a trigram table plus a JavaScript similarity function in place of pg_trgm, a query translator in place of websearch_to_tsquery, and triggers that keep both indexes in step; search_documents gained an integer seq for the FTS rowid. 115 raw SQL sites rewritten across lib, app and queries; lateral joins became window functions. Run reports moved into an automation_reports table. The client is one Drizzle instance over the DB binding; seed and reindex run inside the app as admin jobs and the scripts call them; scripts/export-for-d1.ts copies Supabase into idempotent upserts (types from information_schema, self-referencing tables parents first) and verifies counts and random rows. Migrations live in migrations/ and are applied by wrangler; wrangler.dev.jsonc serves npm run dev without the Response Store Durable Object. Staging runs on its own D1 with the production data imported, search rebuilt, zero exceptions, cold renders 0.4 to 2 s. Docs: ADR-45, DATABASE, ARCHITECTURE, LOCAL-TO-PRODUCTION, CLAUDE.md; CI applies the migrations to a fresh local D1.

+++ src/app/admin/layout.tsx
@@ -14,18 +14,18 @@ async function queueCounts() {
   const [r] = await rawQuery<Record<string, number>>(
     db,
     sql`select
-      (select count(*) from businesses where status = 'pending')::int as businesses,
-      (select count(*) from business_claims where status = 'pending')::int as claims,
-      (select count(*) from professionals where status = 'pending')::int as professionals,
-      (select count(*) from posts where status = 'pending')::int as posts,
-      (select count(*) from reports where status = 'open' and target_type in ('post', 'comment', 'member'))::int as community_reports,
-      ((select count(*) from business_reviews where status = 'pending') + (select count(*) from professional_reviews where status = 'pending'))::int as reviews,
-      (select count(*) from orders where status = 'pending')::int as orders,
-      (select count(*) from submissions where status in ('new', 'reviewing'))::int as submissions,
-      (select count(*) from messages where status = 'new')::int as messages,
-      (select count(*) from inbox_messages where status = 'new' and read_at is null)::int as inbox,
-      (select count(*) from reports where status = 'open')::int as reports,
-      (select count(*) from articles where status = 'draft')::int as drafts`,
+      (select count(*) from businesses where status = 'pending') as businesses,
+      (select count(*) from business_claims where status = 'pending') as claims,
+      (select count(*) from professionals where status = 'pending') as professionals,
+      (select count(*) from posts where status = 'pending') as posts,
+      (select count(*) from reports where status = 'open' and target_type in ('post', 'comment', 'member')) as community_reports,
+      ((select count(*) from business_reviews where status = 'pending') + (select count(*) from professional_reviews where status = 'pending')) as reviews,
+      (select count(*) from orders where status = 'pending') as orders,
+      (select count(*) from submissions where status in ('new', 'reviewing')) as submissions,
+      (select count(*) from messages where status = 'new') as messages,
+      (select count(*) from inbox_messages where status = 'new' and read_at is null) as inbox,
+      (select count(*) from reports where status = 'open') as reports,
+      (select count(*) from articles where status = 'draft') as drafts`,
   );
   return r ?? {};
 }
